+++ src/app/shop-manage/columns.tsx
@@ -3,7 +3,7 @@
 import { Button } from '@/components/ui/button';
 import { ColumnDef } from '@tanstack/react-table';
 import { ArrowUpDown } from 'lucide-react';
-import { DialogRowActions } from '@/components/shop-manage/DiaLogRowActions';
+import { DialogRowActions } from '@/components/shop-manage/dialog-row-actions';
 
 export type Payment = {
   id: string;
